[ET-VK] Add PushConstantDataInfo and vector to hold push constants data in DispatchNode.
Pull Request resolved: https://github.com/pytorch/executorch/pull/7223

This diff adds a new class called `PushConstantDataInfo` to the `DispatchNode` class in the Vulkan backend for Executorch.
This class represents a push constant data entry, which can either be a shared pointer to a tensor's uniform data with an attribute or data with a maximum size of 16 bytes.
The `write` method is also added to this class, which writes the data to a destination buffer.
ghstack-source-id: 256972816
@exported-using-ghexport

Differential Revision: [D66796049](https://our.internmc.facebook.com/intern/diff/D66796049/)